Added HtmlWebpackPlugin. Split Webpack plugins into webpack.plugins.js.
