+++ package.json
@@ -56,6 +56,7 @@
     "eslint-plugin-prettier": "^3.1.4",
     "eslint-plugin-react": "^7.21.3",
     "gh-pages": "^3.1.0",
+    "html-webpack-plugin": "^4.5.0",
     "husky": "^4.3.0",
     "lint-staged": "^10.3.0",
     "node-json-minify": "^1.0.0",
