+++ src/templates/index.template.html
@@ -41,11 +41,9 @@
 
       gtag("config", "UA-178235837-1");
     </script>
-    <title>fusliez notes</title>
   </head>
   <body>
     <noscript>You need to enable Javascript to run this application.</noscript>
     <div id="root"></div>
-    <script src="./index.tsx"></script>
   </body>
 </html>
